services/user_service/response/user_add_city_response.py
from typing import Optional

from pydantic import BaseModel

from .response_status import ResponseStatus


class UserAddCityResponse(BaseModel):
    status: ResponseStatus
    city: Optional[str] = None
