pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^3.4.1"
mypy = "^1.9.0"
pydantic-settings = "^2.2.1"
aio-pika = "^9.4.1"
aiohttp = "^3.9.5"
urllib3 = "^2.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
